requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bleach==6.0.0
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.11
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
defusedxml==0.7.1
diff-match-patch==20230430
discord.py==2.3.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.3
django-import-export==3.2.0
django-markdownify==0.9.3
dulwich==0.21.3
et-xmlfile==1.1.0
filelock==3.12.0
frozenlist==1.4.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.5.0
installer==0.7.0
jaraco.classes==3.2.3
Js2Py==0.71
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Markdown==3.4.3
MarkupPy==1.14
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mutagen==1.46.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
pep517==0.12.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.17
pyjsparser==2.7.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.7
trove-classifiers==2023.4.20
typing_extensions==4.7.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.15
virtualenv==20.21.1
webencodings==0.5.1
websockets==11.0.1
xattr==0.10.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
yt-dlp==2023.3.4
zipp==3.15.0
